mlx5: share CQ entry check
The CQE has owner bit to indicate if it is in SW control or HW.

Share a CQE check for all the mlx5 drivers.
